--- a/src/components/Less.js
+++ b/src/components/Less.js
@@ -27,7 +27,7 @@ class Less extends Preprocessor {
     }
 
     chunkRegex() {
-        return /\.(css|less)$/;
+        return /\.(le|c)ss$/;
     }
 }