This PR improves performance for Sass/Less regex tests

**Sass**
Test String: asdjhjhcjvhxcjhv_askjasldkjaskdj_alksjdalksjdaksj_alksjhdjhfjashdf_xcjvnxcmvnb_,asdasd.scss

| \\.(sa\|sc\|c)ss$ | \\.(css\|s[ac]ss)$ |
| ----------------- | ------------- |
| 1 match (~0ms)  | 1 match, 10 steps (~1ms)  |

**Less**
Test String:
asdjhjhcjvhxcjhv_askjasldkjaskdj_alksjdalksjdaksj_alksjhdjhfjashdf_xcjvnxcmvnb_,asdasd.less

| \\.(le\|c)ss$ | \\.(css\|less)$ |
| ----------------- | ------------- |
| 1 match, 9 steps (~0ms)  | 1 match, 10 steps (~1ms)  |

Testing tool: https://regex101.com/